+++ src/utils.ts
@@ -7,9 +7,8 @@ export interface Alias {
 	// domain of alias
 	domain: string;
 	targetAddress: string;
-	aliasAddress: string;
-	// 0 is disabled, 1 is enabled
-	active: number;
+	address: string;
+	active: boolean;
 	created: Date;
 	modified: Date;
 	// hash of site
@@ -46,8 +45,49 @@ interface FetchAliasData {
 	modified: string | null;
 }
 
+export async function generateAlias(
+	settings: Settings,
+	hostname?: string,
+): Promise<Alias> {
+	// first in msg array should be "alias_added", second is the address, third is the id as a string
+	const { data } = await axios.post<[{ type: string; msg: string[] }]>(
+		`${settings.host}/api/v1/add/alias`,
+		{
+			address: generateEmail(settings, hostname),
+		},
+	);
+
+	if (data[0].type !== "success") {
+		throw new Error("Failed to generate alias");
+	}
+
+	// we need to fetch the alias again to get the creation date because mailcow doesn't return it
+	const alias = await axios.get<FetchAliasData>(
+		`${settings.host}/api/v1/get/alias/${data[0].msg[2]}`,
+		{
+			headers: {
+				"X-API-Key": settings.apiKey,
+			},
+		},
+	);
+
+	return {
+		id: alias.data.id,
+		domain: alias.data.domain,
+		targetAddress: alias.data.goto,
+		address: alias.data.address,
+		active: alias.data.active === alias.data.active_int,
+		// mailcow doesn't return a valid js date for some reason (https://github.com/mailcow/mailcow-dockerized/issues/4876)
+		created: new Date(`${alias.data.created}.000Z`),
+		modified: alias.data.modified
+			? new Date(`${alias.data.modified}.000Z`)
+			: null,
+		siteHash: alias.data.private_comment.split("_")[1],
+	};
+}
+
 export async function fetchAliases(settings: Settings): Promise<Alias[]> {
-	const response = await axios.get<FetchAliasData[]>(
+	const { data } = await axios.get<FetchAliasData[]>(
 		`${settings.host}/api/v1/get/alias/all`,
 		{
 			headers: {
@@ -56,22 +96,20 @@ export async function fetchAliases(settings: Settings): Promise<Alias[]> {
 		},
 	);
 
-	return response.data
+	return data
 		.filter(
 			// make sure alias is active and was generated by this extension
-			(alias) =>
-				alias.active === alias.active_int &&
-				alias.private_comment.startsWith("aliasextension_"),
+			(alias) => alias.private_comment.startsWith("aliasextension_"),
 		)
 		.map((alias) => {
 			return {
 				id: alias.id,
 				domain: alias.domain,
 				targetAddress: alias.goto,
-				aliasAddress: alias.address,
-				active: alias.active,
-				created: new Date(alias.created),
-				modified: alias.modified ? new Date(alias.modified) : null,
+				address: alias.address,
+				active: alias.active === alias.active_int,
+				created: new Date(`${alias.created}.000Z`),
+				modified: alias.modified ? new Date(`${alias.modified}.000Z`) : null,
 				siteHash: alias.private_comment.split("_")[1],
 			};
 		});
@@ -81,7 +119,7 @@ export function generateHash(data: string) {
 	return createHash("sha256").update(data).digest("hex");
 }
 
-export function generateEmail(settings: Settings, hostname: string): string {
+export function generateEmail(settings: Settings, hostname?: string): string {
 	switch (settings.generationMethod) {
 		case GenerationMethod.RandomCharacters:
 			return `${faker.random.alphaNumeric(10)}@${settings.aliasDomain}`;
