+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
   "extends": "plasmo/templates/tsconfig.base",
   "exclude": ["node_modules"],
-  "include": [".plasmo/**/*", "./**/*.ts", "./**/*.tsx"],
+  "include": ["./**/*.ts", "./**/*.tsx"],
   "compilerOptions": {
     "paths": {
-      "~*": ["./src/*.tsx", "./src/*.ts"]
+      "~*": ["./src/*"]
     },
     "baseUrl": "."
   }
